services/requirements.txt
pandas
spanish-dni
fillpdf
selenium
openpyxl
requests
supabase
python-dotenv


